// src/Component/admin-profile/profile/types/ProfileForm.ts
export interface Profile {
  companyName: string;
  address1: string;
  address2: string;
  state: string;
  city: string;
  zip: string;
  yearEstablished: string;
  federalTaxIdNumber: string;
  dunsNumber: string;
  minorityOwned: string;
  paymentTermsId: number;
  services: boolean;
  parts: boolean;
  toolsAndSupplies: boolean;
  oemDealer: boolean;
}